ci: allow canary test to run on custom ceph image
Sometimes we need to run canary test on custom ceph images
to test any issue. For this we always need to create separate
PR with custom image. Now, with github action workflow_dispatch
we can use GitHub UI to pass ceph image as input for CI tests
and that image will be used to run the canary test otherwise,
it will be default value provided.

Signed-off-by: subhamkrai <srai@redhat.com>