test: fix block device detection in github workflow
The machine setup provided by the
CI changed such that the wipefs
command fails not finding the correct
block device to wipe.
This commit fixes the block device "detection"
by adding another possible disk size (75G).
As this is however the same size as the root
disk the "detection" has been adapted to use
the last disk/block device with a matching
size (14G/64G/75G).

Signed-off-by: Jan Klippel <jan.klippel@uhurutec.com>