mgr: handle explicitly the flag AllowPrivilegeEscalation
k8s docs are not clear about the default value of
AllowPrivilegeEscalation, although reading the code suggests it to be
false. To avoid any confusion, let's explicitly initialize the value
of this field. This way, it becomes clear how the container's security
context is initialized, and users can see the current values of the
privileged and AllowPrivilegeEscalation fields when inspecting the
containers.

Fixes: https://github.com/rook/rook/issues/13749

Signed-off-by: Redouane Kachach <rkachach@redhat.com>