test: add integration tests for S3 via keystone
This commit introduces integration tests for s3 and keystone. The tests
are run against the same minimal keystone setup that the tests for swift
and keystone use.

The integration tests use the aws s3 client to use client tooling as
close as possible to the end user instead of using other go-libraries.

Signed-off-by: Silvio Ankermann <silvio.ankermann@cloudandheat.com>