pool: skip crush rule update when not needed
The crush rule will be updated when the failure domain changes.
This update can be skipped when the expected crush rule is
already configured. Otherwise, the log has a confusing message
that makes it appear that the crush rule was updated
when in fact it was not.

Signed-off-by: travisn <tnielsen@redhat.com>
(cherry picked from commit 90bec3120367f71ce9a64b5a121702ed359133f0)