object: set non-ec shared pool to metadata pool
The omap data for multipart uploads is stored in an
additional data pool, which cannot be an EC pool.
So the extra data pool is set to the metadata pool
for object stores with shared pools.

Signed-off-by: Travis Nielsen <tnielsen@redhat.com>
(cherry picked from commit 1c612b0f18c4f19f257ddedfb9cceac0e065e128)