object: virtulhostnames is not required in the endpoint for rgw
The virtualhostnames for rgw is added to possible endpoint list, this
endpoint is mainly used by rook operator for communicating with RGW
server. If enable the virtualhostname feature for rgw, by default the
service endpoint is added to the list. So the service endpoint can
accessed even if the feature is enabled.

Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>